feat: switch to decimal for precision; add FixedPointCalculator
Introduced FixedPointCalculator<T> for high-precision arithmetic.
Replaced double with decimal across the codebase for better accuracy.
Enhanced XLFormulaParser and FunctionExpression to use decimal.
Improved error handling and validation in Parser class.
Updated XLFormulaParser to use FixedPointCalculator<decimal> for
operations like SUM, SUBTRACT, PRODUCT, DIVIDE, etc.
Added new mathematical functions: AVERAGE, COUNT, MAX, MIN, MEDIAN,
STDEV, VAR, ROUND, ABS, POWER, SQRT, LOG, LOG10, CEILING, FLOOR, MOD.
Updated SciCalc project version to 0.1.1.
Improved error messages and exception handling.
Minor formatting and whitespace adjustments in README.md and project
files.

In tests, switched to Convert.ToDouble in CalculatorTests, updated
TestData for more precise expected results, and removed tolerance in
XLFormulaParserTests for exact floating-point comparison.

+++ SciCalc/Sci/XLFormulaParser/NumberExpression.cs
@@ -6,9 +6,9 @@ public partial class Sci
     {
         private class NumberExpression : Expression
         {
-            private readonly double value;
+            private readonly decimal value;
 
-            public NumberExpression(double value)
+            public NumberExpression(decimal value)
             {
                 this.value = value;
             }
